Add a script to delete a list of files from a dataset (#305)
Add a script that deletes all files from a dataset, given a list of datarepo fileIDs. Since a file can't be deleted if a snapshot still refers to it, the script will also delete snapshots as needed before deleting a file.

Add support to `delete_tdr_rows.py` to use this new code, so that snapshots that reference files are automatically deleted if the `delete_files` option is enabled. Also added `dry_run` support to `delete_tdr_rows`